backend/requirements.txt
asgiref==3.8.1
Django==4.2
django-cors-headers==4.6.0
djangorestframework==3.14.0
mysqlclient==2.1.1
Pillow==9.5.0
PyMySQL==1.1.1
python-dotenv==1.0.0
pytz==2024.2
sqlparse==0.5.3
typing_extensions==4.12.2
tzdata==2024.2
